requirements-dev.txt
-r requirements.txt
pytest>=5.4.3
coverage>=5.3
pydantic[email]>=1.7
djangorestframework>=3.10.3
